Fix ramulator readq starvation problem
Bug Issue
- There are three request queues in ramulator
  - readq, writeq, otherq(refresh reqs)

Situation
1. When read requests are inserted & there are also requests in the
otherq, requests from the otherq has priority over the readq. In this
state, the write_mode in Controller.h is set to false.

2. While the reqs in readq is waiting to be scheduled, the program
shifts into the write phase and sends a bunch of write requests into the
ramulator, thus filling the writeq. When this happens, write_mode is set
to true as the number of request in the writeq goes over the highwater
threshold. Since write_mode is set to true, the controller selects requests
from the write queue and schedules them first, thus starving the readq.

Changes
- Changed the readq / writeq selection policy to just select the queue
with the oldest request.